Add native support for Sidekiq
In response to #63, this commit adds reporting of exceptions which happen in Sidekiq processes. The new error_handler API is used so that all exceptions are reported, not just those which occur during job execution:

https://github.com/mperham/sidekiq/blob/master/Upgrading.md#error-service-providers

A server middleware is also added in order to clear the Honeybadger context hash after each job executes. Because earlier versions of Sidekiq had native support for reporting exceptions but did not clear the context hash, the middleware and error handler are installed separately; all Sidekiq versions get the middleware, but only >= 3.0.0 get the error handler.