Fail gracefully when we can't access session data.
When Rails 3 is missing `config.secret_token`, it raises an
ArgumentError when attempting to access the session. This handles that
scenario. Rails 4 raises a RuntimeError before initializing Rails, so
Honeybadger is never invoked (this is better IMO, since important
configuration should be validated before booting the app at all).

Conflicts:
	lib/honeybadger/notice.rb