Be paranoid of Thor
For some reason, when thor is a dependency (but not required) and guard
is required, a `Thor` class is created without actually requiring thor.
This caused and error when attempting to call `Thor.no_commands` (#78).

`Thor.no_commands` was added as of version 0.18.0, so that's the
earliest version we'll support.