Prevent application routes from being cleared in test command.
This fixes #94: The test task creates a custom controller action and
route which is fine when it's not touching any other controller/view
code, but if anything in your Rails stack (such as middleware) includes
route helpers, they are missing when the test action is called.

This bug actually also exists in version 1.x of the gem; I'll eventually
backport this fix to 1.x for Rails 3 and will need to investigate the
same behavior in Rails 2.