Always to_s the api key.
It's possible to be an api key to be all integers, which is a problem
when configuring via ENV because a string of digits is type-cast to
Integer. This just makes sure we compare a string when validating the
api key, and converts the key to a string for use. I'll be adding some
better config validation in the future.