delayed_job improvements.
Added component and action for delayed jobs.

Set context before the delayed job is executed so that all faults reported with Honeybadger.notify have the correct context.

Added support for setting the component and action for delayed_jobs that are not a Delayed::PerformableMethod but simple objects that response to :perform

Include the component & action in the delayed job context so that all messages reported in this job have a proper component & action set.