Filter HTTP_COOKIES in request cgi_data
Previously the HTTP_COOKIE present in the request headers were not being
sanitized (i.e. keys that were present in honey badgers params_filters
were not being filtered.)

This commit introduces a new method to Util::Sanitizer `filter_cookies`
that is responsible for:

1. Parsing the cookies string
2. Sanitizing the parsed cookies string
3. Reconstructing the sanitized cookies.

This method gets called while building the request payload to ensure
that the payload sent to Honeybadger has filtered out any sensitive data
present in the cookies.