Remove source_extract preference.
Previously we would send the template source from a special Rails method
called #source_extract, which is available on
ActionView::Template::Error. This causes a bug when displaying the
source for exceptions which happened outside of the template.  By not
prefering this method, our built-in source extract code will open the
appropriate file (whether it's a .rb or .erb file) and attach the
correct source.