Consolidate require statements.
Originally I'd tried to make each file explicitly require all the
dependencies which it relied on. The benefit of that approach is that
you can require and test each file in isolation. This never happens in
practice, and is cumbersome.

Moving forward, the assumption will be that you must `require
'honeybadger'` (or similar) to use the gem, which allows all of the
supporting files to assume that their dependencies are already required.