Allow before notify hooks to be defined
In order to give other libraries the ability to extend Honeybadger
errors with context outside of what would normally be defined in a call
to `Honeybadger.notify`, we introduce the concept of "before notify"
hooks that accept the notice and allow for modification before it is
sent to the API or queued for later sending.

All before notify hooks are validated for correctness before they are added to
the queue.