Expose the API key via a writer on the notice
With the addition of before notify hooks, we can more intelligently
route notices to different projects via the use of a before hook to set
the API key based on information found in or around the code that raises
a notice. In order to do this, we have to expose the API key as
a property on the notice instead of settings it during initialization
and keeping it through the entire execution of the notice.

An example of this is as follows. The `Ownership` class comes from the
[ownership gem](https://github.com/ankane/ownership).

```ruby
module ErrorHandling
  def self.api_keys
    {
      blog: ENV["HONEYBADGER_BLOG_TEAM_API_KEY"],
      marketing: ENV["HONEYBADGER_MARKETING_TEAM_API_KEY"],
      default: ENV["HONEYBADGER_API_KEY"]
    }
  end
end

Honeybadger.configure do |config|
  config.before_notify do |notice|
    owner = Ownership.owner

    notice.context.merge(tags: owner) if owner
    notice.api_key = ErrorHandling.api_keys.fetch(
      owner,
      ErrorHandling.api_keys[:default]
    )
  end
end
```