Make `Notice#backtrace` assignable.
This is a breaking change.

Before this change, we were exposing the backtrace to callbacks such as
`Honeybadger#exception_filter`, as our internal `Backtrace` object. I
don't think this is ideal, especially now that we want to allow users to
assign/modify the backtrace directly via the new `before_notify` callback.

After this change, `Notice#backtrace` is *always* exposed as the raw
input from one of:

1. Direct user input (i.e. `Honeybadger.notify("message", backtrace: [])`
2. The `#backtrace` from an instance of `Exception`
3. `caller` (default for manual error reports)

It does not parse the backtrace or apply backtrace filters until sending
the notice (after additional user input is impossible).