Fix leak in worker queue
The check for the max queue size is not synchronized, which means it can
return wrong results if multiple threads are pushing on to this queue at
the same time. If the wrong result is returned on the max_queue_size,
the check will be bypassed and the queue will grow unbounded until
memory is exhausted.