Report the right Rails controller and action when using exceptions_app
When Rails is configured to use `exceptions_app` (using its routes), it internally renders a custom controller/action according to `config/routes.rb` (and the response code is 200 ok)

It causes the agent to report the wrong controller name and action, once Rails rendered it.

Right now, as a proof of concept (this can be not the final solution if not according to agent coding practices), checks if `Rails` constant exist and is configure to use exceptions app, if so, resolve the `ORIGINAL_FULLPATH` to the controller/action and use it as new names to be reported.

This solution is more a _guesstimate_ then something 100% assertive (and check every time at runtime), some alternative solutions can be:

* Enable the client to look for routes (user configure)
* On initialization check this condition (at `Honeybadger::Initi::Rails::Railtie`), then use it at runtime, more a performance optimization then assertive

Closes #250