Allow FormattedLogger to accept progname
It appears we pass a FormattedLogger into instances of our
StandardLogger. I made the assumption that StandardLogger is the base
for all loggers (since everything else uses it as a superclass) but
there is an instance that we pass the formatted logger into the
StandardLogger as the delegated logger, phew. To keep things happy, I am
allowing an optional progname to be passed to the Formatted logger and
subsequently, the StandardLogger.