Refactor suspend/shutdown logic
This fixes some confusing logic related to `#shutdown` vs. `#shutdown!`,
and how suspend works.

Suspend allows you to kill the worker and not accept any additional work
until a defined start time.

Shutdown is normally reserved for a user-initiated shutdown, or when
exiting the Ruby process.

Previously, suspending the worker would also put it into a "shutdown"
state, until the start time was reached. This always confused me, as
well as these two similar yet slightly different methods: `#shutdown`
(stop the worker gracefully and stay that way) and `#shutdown!` (kill
the worker, but potentially allow it to be started again in the future).

This change removes the `#shutdown!` method entirely, replacing it with
a private `#kill!` method which is shared by `#shutdown` and `#suspend`.
An argument has been added to the `#shutdown` method which if `true`
will exit immediately without reporting remaining data (to handle the
case where the user just wants a hard stop). "suspended" and "shutdown"
are now two distinct states, with their own predicates.

This also simplifies the `#shutdown` method, removing some cruft that
was leftover from past changes.