Throttling improvements (WIP) (#319)
* Use an integer count to track throttles

* Set a more reasonable throttle backoff

* Reduce the default max_queue_size to 100

* Remove Queue, check size in Worker

* Skip sending data at shutdown when throttled

* Simplify throttle interval calculation

* Improve throttle docs/naming

* Standardize debug helper calls in worker

* Stub queue in spec

This spec group had timing issues which became apparent on JRuby. By
stubbing the size, we don't need to worry about the worker processing
the queue before we can inspect it.

* Fix suspend spec

* Refactor suspend/shutdown logic

This fixes some confusing logic related to `#shutdown` vs. `#shutdown!`,
and how suspend works.

Suspend allows you to kill the worker and not accept any additional work
until a defined start time.

Shutdown is normally reserved for a user-initiated shutdown, or when
exiting the Ruby process.

Previously, suspending the worker would also put it into a "shutdown"
state, until the start time was reached. This always confused me, as
well as these two similar yet slightly different methods: `#shutdown`
(stop the worker gracefully and stay that way) and `#shutdown!` (kill
the worker, but potentially allow it to be started again in the future).

This change removes the `#shutdown!` method entirely, replacing it with
a private `#kill!` method which is shared by `#shutdown` and `#suspend`.
An argument has been added to the `#shutdown` method which if `true`
will exit immediately without reporting remaining data (to handle the
case where the user just wants a hard stop). "suspended" and "shutdown"
are now two distinct states, with their own predicates.

This also simplifies the `#shutdown` method, removing some cruft that
was leftover from past changes.

* Compact

* Fix sprintf order

* Fix some more arguments

* Add id to response debug logging

* Use block-style logging helper for consistency

* Update CHANGELOG.md