Fix spec race condition
I think that the shutdown method was getting past the point of asking if
throttled was true before the worker had recieved the 429 and
incremented the throttle counter. To counter this, I pushed more objects
onto the queue, to give it a little more time to react before calling
shutdown. Not awesome, but so far it seems to have resolved the issue
locally.