Include context in notices from failed Resque jobs (#334)
* Fix resque integration spec

The existing spec was useless, as it would never fail,
even without the code to clear the context.

* Include context in notices from failed Resque jobs

When a developer sets Honeybadger context within the execution of a
Resque job, they expect it to be included in any notices reported from
the job, including when the job fails.

The previous implementation cleared the context in the `around_perform`
hook, which executes before the `on_failure` hook. This meant that the
context was always empty for notices from failed jobs.

This moves clearing the context into the failure handler, and into an
`after_perform` hook. If the job succeeds, the context will be cleared
by `after_perform`. If the job fails, the `after_perform` hook will not run,
so the context will be available when the notice is created in the
`on_failure` hook, which will then clear the context.