Fixed sidekiq retry logic
Our sidekiq plugin relies on `retry_count` in the job parameters to
determine if a job should skip. It was taking the `to_i` of the value
and essentially using it as an attempts count. This was causing the job
to wait 2 extra executions, as the `retry_count` equals 1 on the third
execution (see comment in source for explanation).

This refactor calculates the attempts and uses that to determine if we
should skip reporting. Now if `sidekiq.attempt_threshold` is set to the
value of 2, then on the second execution, it will report (as opposed to
the 4th)