Pass config to Breadcrumbs::Collector when local_context: true
Previously, when using an agent with local_context: true and breadcrumbs
enabled, errors result because when the Breadcrumbs::Collector was
instantiated, the @config instance variable hadn't been initialized
(because opts must be a Hash, not a Config instance). This meant the
collector was passed nil for its config, and when it attempted to access
the breadcrumbs setting, a NoMethodError was raised.

This change ensures that we always set up our agent config before
instantiating the Breadcrumbs::Collector.

I've added a test to check that the collector is passed a config object.
I decided it wasn't particularly important to test the contents of that
object in this specific example, so it's just a type check.

I've used #and_call_original for the stub on the collector instantiation,
despite a double being available to return. I tried returning the double
initially, but there appears to be some state leakage between examples,
as in occasional builds it would complain about the double being present
in other tests.

I briefly looked at tracking down this state leakage, but decided the
scope of changes required to fix it would potentially have significant
knock-on effects, so I've just avoided the possibility of the double
being stored by not using it.