Don't enable Lambda plugin in non-Lambda execution environments
The Lambda plugin currently enables itself in all environments where the
AWS_EXECUTION_ENV environment variable is set, regardless of value.

However, this variable is also set in other contexts, e.g. ECS. In those
contexts this results in the plugin overwriting the component and action
values set in any notice with nil (since no lambda function or handler
data is present in the environment).

This commit updates the requirement test to check for the presence of
the AWS_LAMBDA_FUNCTION_NAME variable instead, which is what gets used
as the component metadata by the plugin.