By default ignore 'ActionDispatch::Http::MimeNegotiation::InvalidType' (#402)
* By default ignore 'ActionDispatch::Http::MimeNegotiation::InvalidType'

If a user-agent sends an Accept or Content-Type header with a mal-formed mime-type, then Rails raises. this happens in my app a lot as a result of what seems to be security vulnerability scanning, people sending Accept headers like `"../../../../../../../../../../etc/passwd{%00`

Rails is not vulnerable, everything is fine, but it raises. And Honeybadger than catches and reports.

Prior to Rals 6.1, the exception class was non-specific so it was hard to ignore/filter. But Rails 6.1 introduces an exception class just for this. Honeybadger should by default ignore it, it fits in with the other Rails stuff Honeybadger ignores by default.

* Update CHANGELOG.md

Co-authored-by: Josh Wood <josh@honeybadger.io>