Accept three arguments into Sidekiq error handler (#495)
Starting with Sidekiq 7.1.5 the error handler is passed three arguments. The new argument is the Sidekiq config. The `_config` key from the second argument has been removed.

Support for passing two arguments and expecting the `_config` is still supported but will raise a deprication warning and will be removed in Sidekiq 8.

This change adds an optional third argument to the error handler lambda defaulted to `Sidekiq.default_configuration` as per the Sidekiq changelog on how to handle this.

This should make the gem forwards compatible with Sidekiq 8 (excluding any other breaking changes).

Sidekiq PR introducing this change: https://github.com/sidekiq/sidekiq/pull/6051