Add Honeybadger::Instrumentation class
This adds a `Honeybadger::Instrumentation` class to help facilitate
gathering metrics within the gem. It's used by calling class methods for
the various types of metrics.

The `Honeybadger::InstrumentationHelper` module can be included into any
class for a cleaner metric DSL.

For this round of implementaiton, the metrics are simply sent off as
events to Insights for collection.