Add a CollectorWorker and collect block
This adds the ability for a plugin to defined a `collect` block, similar
to the `execution` block. During load, these `collect` blocks are
gathered and wrapped in a `CollectionExecution` class and passed to a
queue in a `CollectorWorker`.

The `CollectorWorker` is setup to sychronously execute each collect
block around 1 time per second. A collect block may be configured to
increase it's interval by passing the `interval` option:

```
collect(interval: 10) do
end
```

The above collect block will only be exectued every 10 intervals (about
10 seconds). The `CollectionExecution` class includes the
`Honeybadger::InstrumentationHelper` module, so it has access to any of
the metric methods.

This allows for polling of resources and reporting them to Insights.