feat: add event stats for sidekiq and sq (#638)
Similar to Karafka's stats event, this adds the collected cluster based
stats as a wide event for Sidekiq and SolidQueue. The previous
aggregated metrics can be toggled on, while the new events can be
toggled off.

```
sidekiq:
  insights:
    events: false
solid_queue:
  insights:
    events: false

sidekiq:
  insights:
    metrics: true
solid_queue:
  insights:
    metrics: true
```

In addition, the polling interval has been decreased to 5 seconds to
allow for up to date readings.