feat: update puma and autotuner plugins
This allows the puma and autotuner to emit an event for stats, or
alternavitely aggregate the stats to save on data.

Disable stats events for puma and autotuner:
```
puma:
  insights:
    events: false
autotuner:
  insights:
    events: false
```

Note: The autotuner plugin will still event the `report.autotuner` event
regardless of the above setting.

Enable aggregated metics for puma and autotuner.
```
puma:
  insights:
    metrics: true
autotuner:
  insights:
    metrics: true
```