feat: set context during rails error report (#649)
This will always set the context from Rails when an error is reported
even if it is not handled. This will allow applications to use
`Rails.error.set_context` instead of `Honeybadger.context` and still see
that context in errors in the UI.

It would have been nice if Rails provided an event for when
`set_context` is called so that we can hook into that, but this should
be ok for now.