feat: use after_change for context changes
When an application uses `Rails.error.set_context` to set context, the
gem does not see this data until an error is reported. This change uses
the `ActiveSupport::ExecutionContext.after_change` callback so that we
can always keep track of context changes. This ensures we can report
errors with the proper context outside of Rails error report handling.